+++ src/stapi_fastapi/models/order.py
@@ -29,17 +29,6 @@ OPP = TypeVar("OPP", bound=OpportunityProperties)
 ORP = TypeVar("ORP", bound=OrderParameters)
 
 
-class OrderRequest(BaseModel, Generic[ORP]):
-    datetime: DatetimeInterval
-    geometry: Geometry
-    # TODO: validate the CQL2 filter?
-    filter: CQL2Filter | None = None
-
-    order_parameters: ORP
-
-    model_config = ConfigDict(strict=True)
-
-
 class OrderStatusCode(StrEnum):
     received = "received"
     accepted = "accepted"
@@ -55,6 +44,13 @@ class OrderStatus(BaseModel):
     reason_text: Optional[str] = None
     links: list[Link] = Field(default_factory=list)
 
+    model_config = ConfigDict(extra="allow")
+
+
+class OrderStatuses[T: OrderStatus](BaseModel):
+    statuses: list[T]
+    links: list[Link] = Field(default_factory=list)
+
 
 class OrderSearchParameters(BaseModel):
     datetime: DatetimeInterval
@@ -63,10 +59,10 @@ class OrderSearchParameters(BaseModel):
     filter: CQL2Filter | None = None
 
 
-class OrderProperties(BaseModel):
+class OrderProperties[T: OrderStatus](BaseModel):
     product_id: str
     created: AwareDatetime
-    status: OrderStatus
+    status: T
 
     search_parameters: OrderSearchParameters
     opportunity_properties: dict[str, Any]
@@ -115,3 +111,23 @@ class OrderCollection(_GeoJsonBase):
     def __getitem__(self, index: int) -> Order:
         """get feature at a given index"""
         return self.features[index]
+
+
+class OrderPayload(BaseModel, Generic[ORP]):
+    datetime: DatetimeInterval
+    geometry: Geometry
+    # TODO: validate the CQL2 filter?
+    filter: CQL2Filter | None = None
+
+    order_parameters: ORP
+
+    model_config = ConfigDict(strict=True)
+
+
+class OrderStatusPayload(BaseModel):
+    status_code: OrderStatusCode | None = None
+    reason_code: str | None = None
+    reason_text: str | None = None
+
+    # todo: rework generic types to allow subclasses to be used correctly, and remove extra=allow
+    model_config = ConfigDict(strict=True, extra="allow")